+++ components/Screens/MainPage/MainPage.jsx
@@ -2,7 +2,7 @@ import React, {useState} from "react";
 import { Pressable} from "react-native";
 import { Button } from "react-native";
 
-import { StyledView, StyledText } from "../Style/Style";
+import { StyledView, StyledText } from "../../Style/Style";
 
 const MainPage = ({navigation}) => {
     return (
