+++ components/Screens/MainPage/Map/MapHelperFunctions.jsx
