+++ components/Stack/FindRoomScreenStack.jsx
@@ -1,7 +1,6 @@
 import React from "react"
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import FindRoom from "../Screens/FindRoom";
-
+import FindRoom from "../Screens/MainPage/Map/FindRoom";
 const FindRoomScreenStack = () => {
     ProfileScreen = createNativeStackNavigator();
      return (
