+++ components/Stack/ProfileScreenStack.jsx
@@ -1,7 +1,6 @@
 import React from "react"
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Profile from "../Screens/Profile";
-
+import Profile from "../Screens/MainPage/Profile/Profile";
 const ProfileScreenStack = () => {
     const ProfileScreen = createNativeStackNavigator();
      return (
