+++ components/Stack/SettingPageStack.jsx
@@ -1,6 +1,6 @@
 import React from "react"
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Setting from "../Screens/Setting";
+import Setting from "../Screens/MainPage/Settings/Setting";
 const SettingPageStack = () => {
     const SettingPage = createNativeStackNavigator();
      return (
